pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "usage"
version = "0.1.0"
description = "A simple tool for generating Python project documentation summaries"
readme = "README.md"
license = {text = "CC0"}
authors = [
    {name = "Philip Orange", email = "git@philiporange.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: CC0 1.0 Universal (CC0 1.0) Public Domain Dedication",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
dependencies = [
    "catenator>=0.1.0",
    "orac>=0.1.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
]

[project.scripts]
usage = "usage.cli:cli"

[project.urls]
"Source" = "https://github.com/philiporange/usage"